New version of MI PROCEDURAL
On this new version here some functions has been updated with more flexible usages and also added some more functionalities to make the framework easier. As like the PHP session tasks are able to do easily and also can be permitted by globally. All the other staffs are going to easier to use.